# backend/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "TTBBDD API"
  version: "1.0.0"
  title: "TTBBDD API"
host: "localhost:3000"
tags:
  - name: "characters"
    description: "Charcters API"
  - name: "comics"
    description: "Comics API"
schemes:
  - "http"
paths:
  /characters:
    get:
      tags:
        - "characters"
      summary: "Get all characters"
      description: "Get all characters"
      operationId: "getAll"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findMatchingName:
    get:
      tags:
        - "characters"
      summary: "Finds characters matching name"
      description: ""
      operationId: "findCharactersMatchingName"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Name value that need to be considered for filter"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findByName:
    get:
      tags:
        - "characters"
      summary: "Finds characters by name"
      description: "Multiple name values can be provided with comma separated strings"
      operationId: "findCharactersByName"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Name values that need to be considered for filter"
          required: true
          type: "array"
          items:
            type: "string"
          collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/didYouJustAssumeMyGender:
    get:
      tags:
        - "characters"
      summary: "Finds female characters ordered by amount of powers"
      operationId: "didYouJustAssumeMyGender"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "object"
              properties: {}
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/whoIsPeterPetrelli:
    get:
      tags:
        - "characters"
      summary: "Finds the character with the highest amount of powers"
      operationId: "whoIsPeterPetrelli"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/whoIsRedBarclay:
    get:
      tags:
        - "characters"
      summary: "Finds the fatest character aka Don Solomillón"
      operationId: "whoIsRedBarclay"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findBadHeroes:
    get:
      tags:
        - "characters"
      summary: "Finds the heroes that are villains in one or more crossovers"
      operationId: "findBadHeroes"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findBestInComic:
    get:
      tags:
        - "characters"
      summary: "Finds the best hero in a comic"
      description: ""
      operationId: "findBestInComic"
      produces:
        - "application/json"
      parameters:
        - name: "comicId"
          in: "query"
          description: "Id of the comic"
          required: true
          type: "number"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findSmartestVillain:
    get:
      tags:
        - "characters"
      summary: "Finds the villain (alignment = bad) that has the highest intelligence"
      description: ""
      operationId: "findSmartestVillain"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/findDumbestHero:
    get:
      tags:
        - "characters"
      summary: "Finds the hero (alignment = good) that has the lowest intelligence"
      description: ""
      operationId: "findDumbestHero"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Character"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/count:
    get:
      tags:
        - "characters"
      summary: "Get the count of characters stored"
      description: ""
      operationId: "count"
      produces:
        - "text"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/charactersWithCrossovers:
    get:
      tags:
        - "characters"
      summary: "Get the count of characters that have crossovers"
      description: ""
      operationId: "charsWithCross"
      produces:
        - "text"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/charactersWithInfo:
    get:
      tags:
        - "characters"
      summary: "Get the count of characters that have additional info"
      description: ""
      operationId: "charsWithInfo"
      produces:
        - "text"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/charactersWithPowers:
    get:
      tags:
        - "characters"
      summary: "Get the count of characters that have powers"
      description: ""
      operationId: "charsWithPowers"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/charactersWithStats:
    get:
      tags:
        - "characters"
      summary: "Get the count of characters that have stats"
      description: ""
      operationId: "charsWithStats"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/avgPowersPerCharacter:
    get:
      tags:
        - "characters"
      summary: "Finds the average number of powers per character"
      description: ""
      operationId: "avgPowersPerChar"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/metadata/avgAppearancesPerCharacter:
    get:
      tags:
        - "characters"
      summary: "Finds the average number of appearances per character"
      description: ""
      operationId: "avgAppearancesPerChar"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/ranking/topTenWomen:
    get:
      tags:
        - "characters"
      summary: "Finds the top 10 of women with more appearances in comics"
      description: ""
      operationId: "top10Women"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "object"
              properties: {}
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/ranking/top10Powerfull:
    get:
      tags:
        - "characters"
      summary: "Finds the top 10 of more powerfull characters"
      description: ""
      operationId: "top10Powerfull"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "object"
              properties: {}
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /characters/ranking/top5WithAlignment:
    get:
      tags:
        - "characters"
      summary: "Finds the top 5 of characters with selected alignment and more appearances"
      description: ""
      operationId: "top5CharactersByAlignment"
      produces:
        - "application/json"
      parameters:
        - name: "alignment"
          in: "query"
          description: "Alignment values that need to be considered for filter"
          required: true
          type: "array"
          items:
            type: "string"
            default: "available"
            enum:
              - "good"
              - "bad"
              - "neutral"
          collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "object"
              properties: {}
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Characters"
    x-swagger-router-controller: "Characters"
  /comics/findMatchingName:
    get:
      tags:
        - "comics"
      summary: "Finds comics matching name"
      description: ""
      operationId: "findComicsMatchingName"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Name value that need to be considered for filter"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Comic"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Comics"
    x-swagger-router-controller: "Comics"
  /comics/metadata/count:
    get:
      tags:
        - "comics"
      summary: "Get the count of comics stored"
      description: ""
      operationId: "countComics"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "number"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Comics"
    x-swagger-router-controller: "Comics"
  /comics/findComicWithStrongestHeroes:
    get:
      tags:
        - "comics"
      summary: "Find the comics with the strongest heroes in a sample of 1000 comics"
      description: ""
      operationId: "findComicWithStrongest"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
        "400":
          description: "Invalid status value"
      x-swagger-router-controller: "Comics"
    x-swagger-router-controller: "Comics"
definitions:
  CharacterInfo:
    type: "object"
    properties:
      alignment:
        type: "string"
      gender:
        type: "string"
      eye_color:
        type: "string"
      hair_color:
        type: "string"
      publisher:
        type: "string"
      skin_color:
        type: "string"
      height:
        type: "string"
      weight:
        type: "string"
    example:
      eye_color: "eye_color"
      gender: "gender"
      skin_color: "skin_color"
      publisher: "publisher"
      weight: "weight"
      alignment: "alignment"
      hair_color: "hair_color"
      height: "height"
  CharacterStats:
    type: "object"
    properties:
      intelligence:
        type: "number"
      strength:
        type: "number"
      speed:
        type: "number"
      durability:
        type: "number"
      power:
        type: "number"
      combat:
        type: "number"
    example:
      strength: 1.46581298050294517310021547018550336360931396484375
      durability: 5.63737665663332876420099637471139430999755859375
      combat: 7.061401241503109105224211816675961017608642578125
      power: 2.3021358869347654518833223846741020679473876953125
      speed: 5.962133916683182377482808078639209270477294921875
      intelligence: 6.02745618307040320615897144307382404804229736328125
  CrossOver:
    type: "object"
    properties:
      identity:
        type: "string"
      alignment:
        type: "string"
      status:
        type: "string"
      appearances:
        type: "number"
      firstAppearance:
        type: "string"
      year:
        type: "number"
      universe:
        type: "string"
      extraInfo:
        type: "string"
    example:
      appearances: 3.61607674925191080461672754609026014804840087890625
      firstAppearance: "firstAppearance"
      year: 2.027123023002321833274663731572218239307403564453125
      identity: "identity"
      universe: "universe"
      alignment: "alignment"
      status: "status"
      extraInfo: "extraInfo"
  Character:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      info:
        $ref: "#/definitions/CharacterInfo"
      stats:
        $ref: "#/definitions/CharacterStats"
      comics:
        type: "array"
        items:
          type: "number"
      powers:
        type: "array"
        items:
          type: "string"
      crossOvers:
        type: "array"
        items:
          $ref: "#/definitions/CrossOver"
    example:
      crossOvers:
        - appearances: 3.61607674925191080461672754609026014804840087890625
          firstAppearance: "firstAppearance"
          year: 2.027123023002321833274663731572218239307403564453125
          identity: "identity"
          universe: "universe"
          alignment: "alignment"
          status: "status"
          extraInfo: "extraInfo"
        - appearances: 3.61607674925191080461672754609026014804840087890625
          firstAppearance: "firstAppearance"
          year: 2.027123023002321833274663731572218239307403564453125
          identity: "identity"
          universe: "universe"
          alignment: "alignment"
          status: "status"
          extraInfo: "extraInfo"
      stats:
        strength: 1.46581298050294517310021547018550336360931396484375
        durability: 5.63737665663332876420099637471139430999755859375
        combat: 7.061401241503109105224211816675961017608642578125
        power: 2.3021358869347654518833223846741020679473876953125
        speed: 5.962133916683182377482808078639209270477294921875
        intelligence: 6.02745618307040320615897144307382404804229736328125
      comics:
        - 9.301444243932575517419536481611430644989013671875
        - 9.301444243932575517419536481611430644989013671875
      name: "name"
      _id: 0
      powers:
        - "powers"
        - "powers"
      info:
        eye_color: "eye_color"
        gender: "gender"
        skin_color: "skin_color"
        publisher: "publisher"
        weight: "weight"
        alignment: "alignment"
        hair_color: "hair_color"
        height: "height"
  Comic:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      issue:
        type: "number"
    example:
      issue: 6.02745618307040320615897144307382404804229736328125
      name: "name"
      _id: 0
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"